chg:usr: Fix problem with reading ".../progress" file after aborted.
This is mentioned in (long ago!) issue #80 I've hit the same problem many times, so tracked it down with a fix.

Fix problem with reading ".../progress" file after aborted backup run. This results in an error message (and lack of %backup) when using the --progress command line option.

Note that the code to save the "progress" file worked fine, and this change is so that reading the "progress" file is similar: making the path to the "progress" file, and having the "b" flag for binary i/o.

Fixes #80