fix:usr: make @retry report fatals as exceptions, not log.FatalError.
Final retry exceptions are now passed up to the handler in __main__.py.

This fixes #881 where the outer 'multi' backend actually wants to catch the
exception (onfail=continue) instead of exiting the process.

Based on the preparation in earlier commits, this commit does not affect the
process exit code, it's still backend_error.

Handling of such exceptions in multiprocessing parent/child process is now the
same and no longer requires a special case.

This will enable more cleanups in future commits.